corrected task 1 and added proper tests for it

+++ src/main/java/edu/hw10/task1/generators/ByteFieldGenerator.java
@@ -18,6 +18,6 @@ public class ByteFieldGenerator implements FieldGenerator {
                 max = (byte) ((Max) annotation).value();
             }
         }
-        return new Random().nextInt(min, max);
+        return (byte) new Random().nextInt(min, max);
     }
 }
